Add --private argument
This allows for uploads of packages and index.html files  to your pypi
repository to be marked as private, and therefore ONLY be accessible
via the Cloudfront distribution.  This also negates the need for the
`--secret` argument since these assets will not be available directly
from the S3 bucket, but I decided to leave that argument alone for
backwards compatibility.