extract s3pypi-infrastructure from the s3pypi package
The infrastructure deployment script would not be functional anyway if called as part of a package downloaded from a package repository - it needs access to the CloudFormation templates and SAM expects the lambda function code in a local directory. It's therefore cleaner to keep the infrastructure setup separate from the code that becomes part of the s3pypi package.